refactor(auth) Move onAuthStateChanged to store
