+++ src/services/auth/index.ts
@@ -1,12 +1,9 @@
 /* eslint-disable no-param-reassign */
 import { createAsyncThunk, createSlice, PayloadAction } from '@reduxjs/toolkit';
 import {
-  GoogleAuthProvider, signInWithPopup, signOut,
+  GoogleAuthProvider, signInWithRedirect, signOut,
 } from 'firebase/auth';
-import {
-  addDoc, getDocs, query, where,
-} from 'firebase/firestore';
-import { auth, collections } from 'services/firebase';
+import { auth } from 'services/firebase';
 
 const googleProvider = new GoogleAuthProvider();
 
@@ -37,27 +34,7 @@ const initialState: AuthState = {
 export const authenticateAsync = createAsyncThunk(
   'auth/authenticate',
   async () => {
-    const res = await signInWithPopup(auth, googleProvider);
-    const { user } = res || {};
-    const q = query(collections.users, where('uid', '==', user.uid));
-    const docs = await getDocs(q);
-
-    if (docs.docs.length === 0) {
-      await addDoc(collections.users, {
-        uid: user.uid,
-        name: user.displayName,
-        authProvider: 'google',
-        email: user.email,
-      });
-    }
-
-    return {
-      displayName: user.displayName,
-      email: user.email,
-      emailVerified: user.emailVerified,
-      photoURL: user.photoURL,
-      uid: user.uid,
-    };
+    await signInWithRedirect(auth, googleProvider);
   },
 );
 
@@ -78,9 +55,8 @@ export const authSlice = createSlice({
   // The `extraReducers` field lets the slice handle actions defined elsewhere,
   // including actions generated by createAsyncThunk or in other slices.
   extraReducers(builder) {
-    builder.addCase(authenticateAsync.fulfilled, (state, action) => {
+    builder.addCase(authenticateAsync.fulfilled, (state) => {
       state.isLoadingAuth = false;
-      state.user = action.payload;
     });
     builder.addCase(authenticateAsync.rejected, (state) => {
       state.isLoadingAuth = false;
