+++ src/services/store.ts
@@ -1,9 +1,14 @@
 import { Action, configureStore, ThunkAction } from '@reduxjs/toolkit';
-import auth, { AuthState } from './auth';
+import { onAuthStateChanged } from 'firebase/auth';
+import {
+  addDoc, getDocs, query, where,
+} from 'firebase/firestore';
+import { authSlice, AuthState, setUser } from './auth';
+import { auth, collections } from './firebase';
 
 export const store = configureStore({
   reducer: {
-    auth,
+    auth: authSlice.reducer,
   },
 });
 
@@ -29,3 +34,27 @@ export const authSelectors = {
 export interface AppState {
   auth: AuthState
 }
+
+onAuthStateChanged(auth, async (user) => {
+  if (user) {
+    const q = query(collections.users, where('uid', '==', user.uid));
+    const docs = await getDocs(q);
+
+    if (docs.docs.length === 0) {
+      await addDoc(collections.users, {
+        uid: user.uid,
+        name: user.displayName,
+        authProvider: 'google',
+        email: user.email,
+      });
+    }
+  }
+
+  store.dispatch(setUser(user ? {
+    displayName: user.displayName,
+    email: user.email,
+    emailVerified: user.emailVerified,
+    photoURL: user.photoURL,
+    uid: user.uid,
+  } : undefined));
+});
